Move pkg modules to the root directory
The pkg directory was not necessary, this commit makes navigation
easier.
