+++ main.go
@@ -12,8 +12,8 @@ import (
 	props "github.com/rapid7/cps/api/v1/properties"
 	v2health "github.com/rapid7/cps/api/v2/health"
 	v2props "github.com/rapid7/cps/api/v2/properties"
+	kv "github.com/rapid7/cps/kv"
 	logger "github.com/rapid7/cps/logger"
-	kv "github.com/rapid7/cps/pkg/kv"
 	consul "github.com/rapid7/cps/watchers/v1/consul"
 	file "github.com/rapid7/cps/watchers/v1/file"
 	s3 "github.com/rapid7/cps/watchers/v1/s3"
