+++ watchers/v1/s3/s3.go
@@ -18,8 +18,8 @@ import (
 	"github.com/buger/jsonparser"
 	"go.uber.org/zap"
 
-	"github.com/rapid7/cps/pkg/kv"
-	"github.com/rapid7/cps/pkg/secret"
+	"github.com/rapid7/cps/kv"
+	"github.com/rapid7/cps/secret"
 )
 
 var (
